Added hw configuration for decision procedures.
From reviewer 12:

> The hardware configuration should be specified (in Section 7.1)
> before discussing the runtimes (in Section 7.2)

In section 7.1, we discuss the implementation of the Lucy decision
procedures and runtime. In section 7.2, we evaluate the decision
procedures. In section 7.3, we evaluate the runtime.

At the beginning of section 7.3, we describe the hardware on which we
run the 7.3 experiments. We never described the hardware configuration
of the experiments in 7.2. I think the reviewer is telling us to add it.
This commit does that.