Clarified that decision procedures are run once.
From reviewer 12:

> Experiments in Section 7.2 report some runtimes. The paper seems to
> suggest that identifying invariant confluence in (less than) tenths of
> a second is efficient enough. However, this seems to be relative to
> the targeted throughput of the transaction system (tps).

It took me a while to understand what this reviewer meant. I think that
the reviewer got confused and thought that the decision procedures were
run every single time a transaction was processed. Thus, if the decision
procedures took a 1/10 of a second, the system could only process 10
transactions per second. I added some text to the evaluation section to
clarify that decision procedures are run once _offline_ and never
_online_.