Added CRDT section to related work.
From reviewer 11:

> A clear connection to what CRDTs are within scope would improve the
> contribution of this paper.

and

> You make a connection to Shapiro et al.'s work on CRDT and use a few
> examples that are popular (e.g. PN-Counter), but can you go further
> and connect to more complex structures like Sequence. This will allow
> you to reach a broader audience and let them clearly understand the
> broader implication of this work. (BTW, nice discussion of PK-FK).

Previously, we had a paragraph on CRDTs in the related works section. I
cut it out because we were over the page limit. This commit adds it back
and also expands upon the differences between CRDTs and invariant
confluent objects and why it's natural to combine the two. I also
reference sequence CRDTs, as the reviewer suggests.