Remove optimizations + limitations of segmentation
This commit addresses two pieces of reviewer feedback.

From reviewer 12:

> The optimizations are simply vaguely discussed in Section 6.4.

I completeley removed the vague paragraph on optimizations. It is vague,
not super important, and the paper is a bit long.

Also from reviewer 12:

> Regarding segmented invariant confluence, it does not seem easy to
> find good "quality" segments, which relates to how usable Section 6
> could be. The paper may include some discussions on how segments are
> constructed.

We already had a paragraph saying that our decision procedures can't
help suggest a segmentation, but I agree with the reviewer that we don't
emphasize this enough and don't directly say that the programmer has to
do it themselves. I expanded this paragraph to very explicitly say that
the programmer has to build segmentations and that, yeah, sometimes this
is hard. We leave solving this issue to future work, as it's out of
scope.