First round of cuts.
The paper is now 8 pages without citations. We need 8 pages with
citations, so we need to trim another page or so. I'll probably shrink
the evaluation a bit more, shrink the related works a bit more, shrink
segmented invariant confluence a bit more and see what we end up with.